# Add several shape casts