--- a/src/BoundsUtilities.js
+++ b/src/BoundsUtilities.js
@@ -170,34 +170,38 @@ function boxIntersectsObb( bounds, obbPlanes, obbPoints ) {
 
 }
 
-// TODO: Try this:
 // https://stackoverflow.com/questions/34043955/detect-collision-between-sphere-and-triangle-in-three-js
-
+const closestPointTemp = new Vector3();
+const projectedPointTemp = new Vector3();
+const planeTemp = new Plane();
+const lineTemp = new Line3();
 function sphereIntersectTriangle( sphere, triangle ) {
 
 	const { radius, center } = sphere;
 	const { a, b, c } = triangle;
 
 	// phase 1
-	const line1 = new Line3( a, b );
-	const line2 = new Line3( a, c );
-	const line3 = new Line3( b, c );
-
-	const closestPoint1 = line1.closestPointToPoint( center, true, new Vector3() );
+	lineTemp.start = a;
+	lineTemp.end = b;
+	const closestPoint1 = lineTemp.closestPointToPoint( center, true, closestPointTemp );
 	if ( closestPoint1.distanceTo( center ) <= radius ) return true;
 
-	const closestPoint2 = line2.closestPointToPoint( center, true, new Vector3() );
+	lineTemp.start = a;
+	lineTemp.end = c;
+	const closestPoint2 = lineTemp.closestPointToPoint( center, true, closestPointTemp );
 	if ( closestPoint2.distanceTo( center ) <= radius ) return true;
 
-	const closestPoint3 = line3.closestPointToPoint( center, true, new Vector3() );
+	lineTemp.start = b;
+	lineTemp.end = c;
+	const closestPoint3 = lineTemp.closestPointToPoint( center, true, closestPointTemp );
 	if ( closestPoint3.distanceTo( center ) <= radius ) return true;
 
 	// phase 2
-	const plane = triangle.getPlane( new Plane() );
+	const plane = triangle.getPlane( planeTemp );
 	const dp = Math.abs( plane.distanceToPoint( center ) );
 	if ( dp <= radius ) {
 
-		const pp = plane.projectPoint( center, new Vector3() );
+		const pp = plane.projectPoint( center, projectedPointTemp );
 		const cp = triangle.containsPoint( pp );
 		if ( cp ) return true;