### Updates
- Add a sphere cast
- Add a box cast
- Add a geometry cast
- Update `start` function to build and serve simultaneously
- Add tests for a variety of bounds intersections
- Update MeshBVHVisualizer to display one root
- Add example for shape casts
- Rename `boundsTree.html` to `raycast.html`

### Examples
- [New shapecast example](http://raw.githack.com/gkjohnson/three-mesh-bvh/more-casts/example/bundle/shapecast.html)
- [Old raycast example](http://raw.githack.com/gkjohnson/three-mesh-bvh/more-casts/example/bundle/raycast.html)

### Before Merge
- ~Verify the bounds intersection functions~
- ~#76 Refactor new functions to use a custom shapecast~ Fix #76 
- ~Update Changelog~

### Coming up
- #74
- Performance improvements
    - Look at [this library](https://github.com/juj/MathGeoLib/blob/master/src/Geometry/Triangle.cpp#L708) and Real Time Rendering book for reference
- Update docs
- Add closest distance for shape functions
- Add option to store offset and count on inner nodes so it's easy to just roll up all the triangles in an encapsulated bounds.
- Iterate down both BVH nodes in the geometrycast

### Benchmark
#### before
```
Compute Bounds Tree      : 209.8 ms
Default Raycast          : 26.982143 ms
BVH Raycast              : 0.234779 ms
First Hit Raycast        : 0.051829 ms

BVH Memory Usage         : 10067.068 kb
```

#### after
```
Compute Bounds Tree      : 229.142857 ms
Default Raycast          : 29.339806 ms
BVH Raycast              : 0.144998 ms
First Hit Raycast        : 0.028316 ms
Spherecast               : 0.002838 ms
Boxcast                  : 0.00082 ms
Geometrycast without BVH : 23.880952 ms
Geometrycast with BVH    : 7.491272 ms

BVH Memory Usage         : 10067.068 kb
```

### Shapecast function
```js
function shapecast(intersectsBoundsFunc, intersectsTriangleFunc = null, nodeOrderFunc = null);

// intersectsBoundsFunc(Box3, isLeaf)
// if true is returned then the child bounds are traversed.

// intersectsTriangleFunc(triangle, index0, index1, index2);
// If true is returned then traversal is stopped entirely and the shapecast function will return true.
// If the function is null then no triangles are checked

// nodeOrderFunc(leftNode, rightNode)
// If <= 0 is returned then the left node is traversed first otherwise the right node is
// If the function is null then the left node is traversed first.


```