@gkjohnson 2019-03-04
This pull request **introduces 2 alerts** when merging 006e6539bc2031a1261afa3901ed0c138e4b4469 into d765655c3e55e2022a7d7f45c6358aefbf5e9053 - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-04c3e79100bdd1ba3422ad87b88d13f23f9891ce)

**new alerts:**

* 2 for Unused variable, import, function or class

---

*Comment posted by [LGTM.com](https://lgtm.com)*