@gkjohnson 2019-03-06
This pull request **introduces 2 alerts** when merging 385b57ffbdca900b1ae96c10b6ee8b484e2cecd8 into d765655c3e55e2022a7d7f45c6358aefbf5e9053 - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-9eea7caa589ebb26311eeafb363627b9b7b17f8c)

**new alerts:**

* 2 for Assignment to constant

---

*Comment posted by [LGTM.com](https://lgtm.com)*