@gkjohnson 2019-03-06
This pull request **introduces 2 alerts** when merging 1a88f2387d19f4417394869a3c699339cc09535d into d765655c3e55e2022a7d7f45c6358aefbf5e9053 - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-51d88b52ce00492d8de0111f9b46ea2177d4dc40)

**new alerts:**

* 2 for Assignment to constant

---

*Comment posted by [LGTM.com](https://lgtm.com)*