@gkjohnson 2019-03-06
This pull request **introduces 2 alerts** when merging 409731c1e4ffca35079441c36aecd94c530de89b into d765655c3e55e2022a7d7f45c6358aefbf5e9053 - [view on LGTM.com](https://lgtm.com/projects/g/gkjohnson/three-mesh-bvh/rev/pr-8ca514943171c0d32201812413e892b09d6cc6bb)

**new alerts:**

* 2 for Assignment to constant

---

*Comment posted by [LGTM.com](https://lgtm.com)*